BUG: fix garbage data when http-send-name-header replaces an existing header
This patch is an attempt to prevent sending garbage data when
http-send-name-header replaced existing headers in the request.

http-send-name-header is applied late in the request processing. The buffer is
already ready to be sent to the backend server. When headers are removed, the
data length is not modified, resulting in sending more data than required. By
reducing the data length to send after removing them, this should fix the
issue.

This patch doesn't need to be ported to haproxy-1.5, which already readjust the
amount of data to be forwarded.

This issue was reported by Michael Seiferle at BaseX, who also tested and
confirmed the fix.